restructure of data utils
splitted into destined files (glyph_utils, conversions etc
