+++ setup.py
@@ -5,5 +5,6 @@ setup(
     version='0.1',
     install_requires=['ufo-extractor', 'ufo2ft', 'arabic-reshaper',
                       'fontTools', 'defcon', 'svgpath2mpl', 'python-bidi',
+                      'tqdm', 
                       'pysilfont @ git+https://github.com/silnrsi/pysilfont'],
 )
